cleaned up penalties plugin, added collision penalties and restructured a bit. Also fix collision speed on live timings

+++ config_ini.go
@@ -427,7 +427,7 @@ type CurrentRaceConfig struct {
 
 	CollisionPenaltiesEnabled             bool                 `ini:"-"`
 	CollisionPenaltiesIgnoreFirstLap      bool                 `ini:"-"`
-	CollisionPenaltiesOnlyOverSpeed       int                  `ini:"-"`
+	CollisionPenaltiesOnlyOverSpeed       float32              `ini:"-"`
 	CollisionPenaltiesNumWarnings         int                  `ini:"-"`
 	CollisionPenaltiesPenaltyType         acserver.PenaltyType `ini:"-"`
 	CollisionPenaltiesBoPAmount           float32              `ini:"-"`
